refactor(components): 파일 구조 정리 / DP-96
- 컴포넌트 파일 구조 정리 진행
