+++ src/components/organisms/Sidebar/MainSidebar/MainSidebar.module.scss
