+++ src/layouts/MainLayout/MainLayout.tsx
@@ -2,8 +2,8 @@ import styles from './MainLayout.module.scss';
 
 import { Outlet } from '@tanstack/react-router';
 
-import Header from '@/components/Header/Header';
-import { Sidebar } from '@/components/Sidebar/MainSidebar/MainSidebar';
+import Header from '@/components/organisms/Header/Header';
+import { Sidebar } from '@/components/organisms/Sidebar/MainSidebar/MainSidebar';
 
 const MainLayout = () => {
   return (
