+++ src/pages/Auth/SignUpPage/SignUpForm.tsx
@@ -7,7 +7,7 @@ import { useNavigate } from '@tanstack/react-router';
 import Input from '@/components/atoms/Input/Input';
 import PasswordInput from '@/components/atoms/Input/PasswordInput';
 import Button from '@/components/atoms/Button/Button';
-import FormField from '@/components/molecules/FormField';
+import FormField from '@/components/molecules/FormField/FormField';
 import styles from './SignUpPage.module.scss';
 
 export default function SignUpForm() {
