+++ src/pages/Main/SharedRepositoryPage/SharedRepositoriesPage.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 
 import Toggle from '@/components/atoms/Toggle/Toggle';
-import Pagination from '@/components/Pagination/Pagination';
-import Repository from '@/components/Repository/Repository';
+import Pagination from '@/components/molecules/Pagination/Pagination';
+import Repository from '@/components/organisms/RepoListItem/RepoListItem';
 
 import styles from './SharedRepositoriesPage.module.scss';
